+++ lib/core/notification_handler.dart
@@ -45,6 +45,76 @@ class NotificationHandler {
     if (!notificationService.existsSync()) {
       notificationService.createSync();
       notificationService.writeAsStringSync('''
+import 'dart:convert';
+
+import 'package:firebase_messaging/firebase_messaging.dart';
+import 'package:flutter_local_notifications/flutter_local_notifications.dart';
+import 'package:$projectName/main.dart';
+import 'package:http/http.dart' as http;
+
+class NotificationService {
+
+  // Api function to send push messages via the app itself
+  static sendPushMessage(
+      {required String token,
+      required String body,
+      required String title}) async {
+    try {
+      await http.post(Uri.parse('https://fcm.googleapis.com/fcm/send'),
+          headers: {
+            'Content-Type': 'application/json',
+            'Authorization':
+                'key=<FIREBASE_CLOUD_MESSAGING_KEY>'
+          },
+          body: jsonEncode(<String, dynamic>{
+            'priority': 'high',
+            'data': <String, dynamic>{
+              'click_action': 'FLUTTER_NOTIFICATION_CLICK',
+              'status': 'done',
+              'body': body,
+              'title': title,
+            },
+            "notification": <String, dynamic>{
+              "title": title,
+              "body": body,
+              "android_channel_id": "<CHANNEL_NAME>"
+            },
+            "to": token
+          }));
+    } catch (e) {
+      print(e.toString());
+    }
+  }
+
+  static Future<void> initLocalNotifications() async {
+    const AndroidInitializationSettings initializationSettingsAndroid =
+        AndroidInitializationSettings('@mipmap/ic_launcher');
+    final InitializationSettings initializationSettings =
+        InitializationSettings(android: initializationSettingsAndroid);
+    await flutterLocalNotificationsPlugin.initialize(initializationSettings);
+  }
+
+  static Future<void> showLocalNotification(RemoteMessage message) async {
+    const AndroidNotificationDetails androidPlatformChannelSpecifics =
+        AndroidNotificationDetails(
+      '<channel_name>', // Change this to a unique channel ID
+      'Your Channel Name', // no need to change this
+      channelDescription: '',
+      importance: Importance.max,
+      priority: Priority.high,
+    );
+    const NotificationDetails platformChannelSpecifics =
+        NotificationDetails(android: androidPlatformChannelSpecifics);
+
+    await flutterLocalNotificationsPlugin.show(
+      0,
+      message.data['title'],
+      message.data['body'],
+      platformChannelSpecifics,
+      payload: 'item x',
+    );
+  }
+}
 
 ''');
     }
@@ -99,5 +169,6 @@ class NotificationHandler {
 </manifest>
 
 ''');
+
   }
 }