+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.3
+
+- Fix: Notification Services 
+
+## 1.0.2
+
+- Fix: main.dart when no auth service is selected
+
 ## 1.0.1
 
 - Added menu driven flow